requirements.txt
ansible>=2.9.0
netaddr>=0.8.0
molecule[docker]
pytest-testinfra
